move to PropRef when updating graph, avoid str clone
