+++ db4-storage/src/pages/node_page/writer.rs
@@ -10,7 +10,7 @@ use raphtory_api::core::entities::{
     EID, GID, VID,
     properties::{
         meta::{NODE_ID_IDX, NODE_TYPE_IDX, STATIC_GRAPH_LAYER_ID},
-        prop::{AsPropRef, Prop},
+        prop::{AsPropRef, Prop, PropRef},
     },
 };
 use raphtory_core::{
@@ -160,20 +160,20 @@ impl<'a, MP: DerefMut<Target = MemNodeSegment> + 'a, NS: NodeSegmentOps> NodeWri
         }
     }
 
-    pub fn check_metadata(
+    pub fn check_metadata<P: AsPropRef>(
         &self,
         pos: LocalPOS,
         layer_id: usize,
-        props: &[(usize, Prop)],
+        props: &[(usize, P)],
     ) -> Result<(), StorageError> {
         self.mut_segment.check_metadata(pos, layer_id, props)
     }
 
-    pub fn update_c_props(
+    pub fn update_c_props<P: AsPropRef>(
         &mut self,
         pos: LocalPOS,
         layer_id: usize,
-        props: impl IntoIterator<Item = (usize, Prop)>,
+        props: impl IntoIterator<Item = (usize, P)>,
     ) {
         let (is_new_node, add) = self.mut_segment.update_metadata(pos, layer_id, props);
         self.mut_segment.increment_est_size(add);
