+++ db4-storage/src/segments/graph_prop/segment.rs
@@ -3,7 +3,10 @@ use crate::{
     segments::{HasRow, SegmentContainer},
     wal::LSN,
 };
-use raphtory_api::core::entities::properties::{meta::Meta, prop::Prop};
+use raphtory_api::core::entities::properties::{
+    meta::Meta,
+    prop::{AsPropRef, Prop, PropRef},
+};
 use raphtory_core::{
     entities::properties::tprop::TPropCell,
     storage::timeindex::{AsTime, EventTime},
@@ -99,10 +102,10 @@ impl MemGraphPropSegment {
         self.lsn = lsn;
     }
 
-    pub fn add_properties<T: AsTime>(
+    pub fn add_properties<T: AsTime, P: AsPropRef>(
         &mut self,
         t: T,
-        props: impl IntoIterator<Item = (usize, Prop)>,
+        props: impl IntoIterator<Item = (usize, P)>,
     ) -> usize {
         let layer = self.get_or_create_layer(Self::DEFAULT_LAYER);
         let est_size = layer.est_size();
@@ -116,7 +119,7 @@ impl MemGraphPropSegment {
         layer_est_size - est_size
     }
 
-    pub fn check_metadata(&self, props: &[(usize, Prop)]) -> Result<(), StorageError> {
+    pub fn check_metadata<P: AsPropRef>(&self, props: &[(usize, P)]) -> Result<(), StorageError> {
         if let Some(layer) = self.layers.get(Self::DEFAULT_LAYER) {
             layer.check_metadata(Self::DEFAULT_ROW.into(), props)?;
         }
@@ -124,7 +127,10 @@ impl MemGraphPropSegment {
         Ok(())
     }
 
-    pub fn update_metadata(&mut self, props: impl IntoIterator<Item = (usize, Prop)>) -> usize {
+    pub fn update_metadata<P: AsPropRef>(
+        &mut self,
+        props: impl IntoIterator<Item = (usize, P)>,
+    ) -> usize {
         let segment_container = self.get_or_create_layer(Self::DEFAULT_LAYER);
         let est_size = segment_container.est_size();
