+++ raphtory-api/src/core/entities/properties/prop/mod.rs
@@ -1,6 +1,8 @@
 pub mod arrow;
 
 mod prop_array;
+
+pub mod prop_col;
 mod prop_enum;
 mod prop_ref_enum;
 mod prop_type;
